+++ src/main/java/com/example/examplemod/ExampleModDataGen.java
@@ -7,6 +7,7 @@ import java.util.concurrent.CompletableFuture;
 import com.example.examplemod.datagen.ModBlockLootTableProvider;
 import com.example.examplemod.datagen.ModBlockTagsProvider;
 import com.example.examplemod.datagen.ModDataMapProvider;
+import com.example.examplemod.datagen.ModEquipmentAssetProvider;
 import com.example.examplemod.datagen.ModItemTagsProvider;
 import com.example.examplemod.datagen.ModModelProvider;
 import com.example.examplemod.datagen.ModRecipeProvider;
@@ -38,5 +39,6 @@ public class ExampleModDataGen {
         generator.addProvider(true, new ModRecipeProvider.Runner(packOutput, lookupProvider));
         generator.addProvider(true, new ModDataMapProvider(packOutput, lookupProvider));
         generator.addProvider(true, new ModItemTagsProvider(packOutput, lookupProvider));
+        generator.addProvider(true, new ModEquipmentAssetProvider(packOutput));
     }
 }